+++ templates/parts/admin/content-starter-excerpt.php
@@ -5,7 +5,8 @@
  * @package    Michelle
  * @copyright  WebMan Design, Oliver Juhas
  *
- * @since  1.0.0
+ * @since    1.0.0
+ * @version  1.0.4
  */
 
 // Exit if accessed directly.
@@ -18,6 +19,6 @@ if ( ! is_customize_preview() ) {
 ?>
 
 <p>
-	<a href="https://themedemos.webmandesign.eu/michelle/page-excerpt/"><?php echo esc_html_x( 'This is a page/post excerpt text.', 'Theme starter content', 'michelle' ); ?></a>
+	<a href="https://webmandesign.github.io/docs/michelle/#posts-excerpt"><?php echo esc_html_x( 'This is a page/post excerpt text.', 'Theme starter content', 'michelle' ); ?></a>
 	Lorem ipsum dolor sit amet consectetur adipiscing elit gravida, enim curabitur sagittis magna molestie rutrum nostra viverra, libero posuere commodo class consequat iaculis.
 </p>